feat: merged init-poap.ts into poap.ts
jira: DCD-53
