+++ utils/poap.ts
@@ -1,9 +1,9 @@
 import {DesmosClient, OfflineSignerAdapter, SigningMode} from "@desmoslabs/desmjs";
 import {program} from "commander";
 import * as Config from "./config"
-import {EventInfo, ExecuteMsg, QueryMsg} from "@desmoslabs/contract-types/contracts/poap";
+import {EventInfo, ExecuteMsg, InstantiateMsg, QueryMsg} from "@desmoslabs/contract-types/contracts/poap";
 import {AccountData} from "@cosmjs/amino";
-import {parseBool, parseCWTimestamp} from "./cli-parsing-utils";
+import {parseBool, parseCWTimestamp, parseIpfsUri} from "./cli-parsing-utils";
 
 async function main() {
     const signer = await OfflineSignerAdapter.fromMnemonic(SigningMode.DIRECT, Config.mnemonic);
@@ -14,6 +14,45 @@ async function main() {
 
     program.name("Poap contract utils")
         .description("Utility script to interact with the poap contract");
+
+    program.command("init")
+        .description("Initialize a new instance of a poap contract")
+        .requiredOption("--code-id <code-id>", "Id of the contract to initialize", parseInt)
+        .requiredOption("--cw721-code-id <cw721-code-id>", "Id of the cw721 contract that will be initialized from the poap contract", parseInt)
+        .requiredOption("--name <name>", "Poap name")
+        .requiredOption("--symbol <symbol>", "Poap symbol")
+        .requiredOption("--start <start>", "Datetime when the event will start in RFC3339 format (2022-12-31:10:00:00)", parseCWTimestamp)
+        .requiredOption("--end <end>", "Date when the event will end in RFC3339 format (2022-12-31:10:00:00)", parseCWTimestamp)
+        .requiredOption("--address-limit <address_limit>", "Max number of poap that an user can receive", parseInt)
+        .requiredOption("--poap-uri <poap_uri>", "ipfs poap uri that contains the poap metadata", parseIpfsUri)
+        .requiredOption("--event-uri <event_uri>", "ipfs event uri that contains the event metadata", parseIpfsUri)
+        .option("--creator <creator>", "Bech32 address of who is creating the event", account!.address)
+        .option("--minter <minter>", "Bech32 address of who will have the minting rights", account!.address)
+        .option("--admin <admin>", "Bech32 address of who will have the contract admin rights", account!.address)
+        .action(async (options) => {
+            console.log(`Initializing contract with code ${options.codeId}`);
+            let instantiateMsg: InstantiateMsg = {
+                admin: options.admin,
+                minter: options.minter,
+                cw721_code_id: options.cw721CodeId.toString(),
+                event_info: {
+                    creator: options.creator,
+                    base_poap_uri: options.poapUri.toString(),
+                    event_uri: options.eventUri.toString(),
+                    start_time: options.start,
+                    end_time: options.end,
+                    per_address_limit: options.addressLimit
+                },
+                cw721_initiate_msg: {
+                    name: options.name,
+                    minter: options.minter,
+                    symbol: options.symbol
+                }
+            }
+
+            const initResult = await client.instantiate(account!.address, options.codeId, instantiateMsg, options.name, "auto");
+            console.log("Contract initialized", initResult);
+        })
 
     program.command("query")
         .description("Queries the contract state")
@@ -80,6 +119,7 @@ async function main() {
                 console.log("Nothing to update");
                 return;
             }
+
             const currentEventInfo: EventInfo = await client.queryContractSmart(options.contract, { event_info: {} } as QueryMsg);
             let msg: ExecuteMsg = {
                 update_event_info: {
